requirements.txt
playsound
openai
openai-whisper
python-dotenv
PyAudio
PyObjC
google-cloud-texttospeech
